Change 'make serve' so that it opens a browser too
Not sure how ideal this is. It's kind of nice, but probably in the
longer run, it would be best to combine these few commands in the
Makefile into a script. The flaw here is that to kill the server, you
have to find the process id and `kill $PID`.